srcs/BackEnd/project/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.2
django-sslserver==0.22
gunicorn==21.2.0
idna==3.6
packaging==23.2
psycopg==3.1.18
psycopg-binary==3.1.18
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.2.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
pytz==2024.1
